+++ src/orchestration/ChainSharp.Effect.Orchestration.Scheduler/Workflows/TaskServerExecutor/ExecuteManifestRequest.cs
@@ -8,4 +8,4 @@ namespace ChainSharp.Effect.Orchestration.Scheduler.Workflows.TaskServerExecutor
 /// Optional workflow input object for ad-hoc executions (e.g., from the dashboard).
 /// When null, the input is resolved from the associated Manifest's properties.
 /// </param>
-public record ExecuteManifestRequest(int MetadataId, object? Input = null);
+public record ExecuteManifestRequest(long MetadataId, object? Input = null);
